apply devalue factor and fiddle with it

+++ backend/migrations/2021-09-05-032727_initial/up.sql
@@ -106,7 +106,7 @@ CREATE TABLE collections (
   notoriety_type TEXT NOT NULL,
   notoriety_score REAL NOT NULL,
   notoriety_score_explanation TEXT NOT NULL,
-  harvest_time_devalue_factor DOUBLE,
+  harvest_time_devalue_factor REAL,
 
   title TEXT,
   author TEXT,
